requirements.txt
mysql_connector_repackaged==0.3.1
python-dotenv==1.1.0
python-telegram-bot==22.0
